requirements/base.txt
# Python
# ------------------------------------------------------------------------------
argon2-cffi  # https://github.com/hynek/argon2_cffi
python-slugify  # https://github.com/un33k/python-slugify
Pillow  # https://github.com/python-pillow/Pillow
pytz  # https://github.com/stub42/pytz
redis  # https://github.com/antirez/redis
whitenoise  # https://github.com/evansd/whitenoise
dj-database-url


# Django
# ------------------------------------------------------------------------------
django
django-allauth
django-environ  # https://github.com/joke2k/django-environ
django-redis  # https://github.com/niwinz/django-redis
django-taggit  # https://github.com/alex/django-taggit
sorl-thumbnail  # https://github.com/jazzband/sorl-thumbnail
django-debug-toolbar  # https://github.com/jazzband/django-debug-toolbar
django-extensions  # https://github.com/django-extensions/django-extensions
django-crispy-forms  # https://github.com/django-crispy-forms/django-crispy-forms

# Django REST
# ------------------------------------------------------------------------------
djangorestframework
django-rest-auth


# Angular
# ------------------------------------------------------------------------------
django-cors-headers
django-material-admin
django-angular

# Channels
# ------------------------------------------------------------------------------
channels
channels-redis  # https://github.com/django/channels_redis
raven
environ
daphne

# Scrape
# ------------------------------------------------------------------------------
beautifulsoup4
soupsieve